useless static_cast

+++ src/search16.cc
@@ -508,7 +508,7 @@ auto backtrack_16(char * qseq,
       const uint64_t d
         = dirbuffer[(offset
                      + longestdbsequence * 4 * static_cast<uint64_t>(j / 4)
-                     + static_cast<uint64_t>(4 * i + (static_cast<uint64_t>(j) & 3U)))
+                     + static_cast<uint64_t>(4 * i + (j & 3U)))
                     % dirbuffersize];  // refactoring: how to rename that variable?
 
       if ((operation == 'I') and ((d & maskextleft) == 0U))
